# HEARTH/contracts/openapi.json
{"openapi":"3.1.0","info":{"title":"HEARTH edge research API","version":"0.1.0"},"paths":{"/healthz":{"get":{"summary":"Healthz","operationId":"healthz_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/openapi.json":{"get":{"summary":"Openapi","operationId":"openapi_v1_openapi_json_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Openapi V1 Openapi Json Get"}}}}},"security":[{"APIKeyHeader":[]}]}},"/v1/batches":{"post":{"summary":"Register","operationId":"register_v1_batches_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Batch"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Register V1 Batches Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/v1/readings:batch":{"post":{"summary":"Readings","operationId":"readings_v1_readings_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadingBatch"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Readings V1 Readings Batch Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/v1/stalls/{stall_id}/state":{"get":{"summary":"Stall State","operationId":"stall_state_v1_stalls__stall_id__state_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"stall_id","in":"path","required":true,"schema":{"type":"string","title":"Stall Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Stall State V1 Stalls  Stall Id  State Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/stalls/{stall_id}/batches/{batch_id}/shelf-life":{"get":{"summary":"Shelf Life","operationId":"shelf_life_v1_stalls__stall_id__batches__batch_id__shelf_life_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"stall_id","in":"path","required":true,"schema":{"type":"string","title":"Stall Id"}},{"name":"batch_id","in":"path","required":true,"schema":{"type":"string","title":"Batch Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Shelf Life V1 Stalls  Stall Id  Batches  Batch Id  Shelf Life Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/batches/{batch_id}:close":{"post":{"summary":"Close","operationId":"close_v1_batches__batch_id__close_post","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"batch_id","in":"path","required":true,"schema":{"type":"string","title":"Batch Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CloseRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Close V1 Batches  Batch Id  Close Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ledger/events":{"get":{"summary":"Events","operationId":"events_v1_ledger_events_get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"after","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"After"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":1000,"minimum":1,"default":100,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Events V1 Ledger Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/ledger/verify":{"get":{"summary":"Verify","operationId":"verify_v1_ledger_verify_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Verify V1 Ledger Verify Get"}}}}},"security":[{"APIKeyHeader":[]}]}},"/v1/demo/run":{"post":{"summary":"Demo","operationId":"demo_v1_demo_run_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Demo V1 Demo Run Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"Batch":{"properties":{"batch_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Batch Id"},"market_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Market Id"},"stall_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Stall Id"},"product_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Product Id"},"device_ids":{"items":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$"},"type":"array","maxItems":64,"minItems":1,"title":"Device Ids"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires At"},"initial_quality":{"type":"number","maximum":1.0,"minimum":0.0,"title":"Initial Quality","default":1.0}},"additionalProperties":false,"type":"object","required":["batch_id","market_id","stall_id","product_id","device_ids","created_at"],"title":"Batch"},"CloseRequest":{"properties":{"reason":{"type":"string","maxLength":500,"minLength":3,"title":"Reason"}},"additionalProperties":false,"type":"object","required":["reason"],"title":"CloseRequest"},"DemoRequest":{"properties":{"scenario":{"type":"string","enum":["normal","cooled","heatwave","outage","sensor_fault"],"title":"Scenario","default":"normal"},"hours":{"type":"number","maximum":72.0,"minimum":1.0,"title":"Hours","default":24},"interval_minutes":{"type":"integer","maximum":60.0,"minimum":1.0,"title":"Interval Minutes","default":10},"seed":{"type":"integer","maximum":4294967295.0,"minimum":0.0,"title":"Seed","default":20260916}},"additionalProperties":false,"type":"object","title":"DemoRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"Metric":{"properties":{"metric":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Metric"},"value":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Value"},"unit":{"type":"string","maxLength":32,"minLength":1,"title":"Unit"},"quality":{"type":"string","enum":["ok","missing","fault","saturated","uncalibrated"],"title":"Quality","default":"ok"}},"additionalProperties":false,"type":"object","required":["metric","value","unit"],"title":"Metric"},"Reading":{"properties":{"schema_version":{"type":"string","const":"1.0","title":"Schema Version","default":"1.0"},"event_id":{"type":"string","format":"uuid","title":"Event Id"},"boot_id":{"type":"string","format":"uuid","title":"Boot Id"},"sequence":{"type":"integer","minimum":0.0,"title":"Sequence"},"market_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Market Id"},"stall_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Stall Id"},"batch_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Batch Id"},"device_id":{"type":"string","pattern":"^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$","title":"Device Id"},"observed_at":{"type":"string","format":"date-time","title":"Observed At"},"source":{"type":"string","enum":["synthetic","sensor"],"title":"Source"},"metrics":{"items":{"$ref":"#/components/schemas/Metric"},"type":"array","maxItems":32,"minItems":1,"title":"Metrics"}},"additionalProperties":false,"type":"object","required":["boot_id","sequence","market_id","stall_id","batch_id","device_id","observed_at","source","metrics"],"title":"Reading"},"ReadingBatch":{"properties":{"readings":{"items":{"$ref":"#/components/schemas/Reading"},"type":"array","maxItems":256,"minItems":1,"title":"Readings"}},"additionalProperties":false,"type":"object","required":["readings"],"title":"ReadingBatch"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"X-API-Key"}}}}